Seperate tests and handle etcd3
